reorganise SFRRT functions
